chore: update `deny.toml`.
Updates `deny.toml` to remove an unused license warning and to
add an ignore for the `instant` dependency advisory as the
dependency is only activated for `wasm32*` targets, which are not
supported for the `wdl` family of crates currently.

Additionally, the `instant` dependency will be removed upon the
next release of `reqwest-retry`.