fix: add Docker backend to engine tests.
This commit adds the Docker backend to the engine tests so both the
task and workflow tests run with the local and docker backends.

Fixed a bug in the type definition of the `container` field of the
task variable, which should be optional.

Updated `crankshaft` to latest commit for environment variable
support.

Fixed a few bugs in path translation:

* Recursive placeholder evaluation was only translating paths at
  the top most placeholder; the fix was to let the evaluation
  context do the path translation so that it can be accessed from
  any part of expression evaluation.
* The temporary directories weren't being inserted into the path
  trie, which meant they were not included for commands with
  placeholders that created files directly.