refactor: refactor `wdl-ast` to allow for alternative tree representation.
This commit refactors `wdl-ast` so that all of the AST types take
a type parameter for either a node or a token.

This change allows for the AST to operate off of a different
representation of `SyntaxNode` and `SyntaxToken` for evaluation.

In the case of `wdl-engine`, evaluation is performed asynchronously
and so the node and token types must be `Send`; however, the
implementation from `rowan` is not `Send`. With this refactoring,
the engine can "morph" the AST returned from analysis to an
AST with its own implementation of `SyntaxNode` and `SyntaxToken`
that are `Send`.

This cleans up quite a few things in the engine, most notably the
engine no longer has to rebuild task and workflow evaluation graphs
with `AstPtr`-based graph nodes.

It will also allow the engine's expression evaluator to become
fully async in an upcoming change to allow for the stdlib functions
to be asynchronous, which is require for the file functions that
may fetch files from remote storage.

Many of the changes here stem from a decision to rename two methods
in `wdl-ast`: `inner` instead of `syntax` and `text` instead of
`as_str`. The rational for using `inner` is that since the underlying
types are generic and do not have to be `SyntaxNode` or `SyntaxToken`,
calling the method `syntax` no longer makes sense. Additionally, we
need a method on the abstract node type for getting its text and
`as_str` doesn't work since it must return `impl fmt::Display`;
therefore, I decided to name it `text` on both the node and token
traits.