refactor: make expression evaluation async.
This commit changes the expression evaluator in `wdl-engine` to be
async.

This is required for an upcoming change to function evaluation
that will allow for the `read_*` functions to work from remote
files.

Note that this change does not include a change to the stdlib
functions to make their calls actually async; that change will be
made shortly in a follow-up commit.