feat: allow reading of remote files from the stdlib file functions.
This commit implements reading of remote files via the `read_*`
functions in the WDL standard library.

Currently the support is limited to http/https URLs, but in the near
future this will be extended to support Azure Storage, S3, and
Google Cloud Storage URLs (along with presigned auth token support).

This commit does not implement proper input localization for the
Docker backend, but that implementation will also be completed soon.

Allowing the `read_*` functions to operate off of URLs will enable
remote task execution backends like TES to only need to fetch
files out of cloud storage if the local workflow evaluation attempts
to read them.