feat: support cloud storage URIs in the downloader.
This commit implements support for cloud storage URIs:

* `az://<account>/<container>/<blob>` for Azure Blob Storage
* `s3://<bucket>/<object>` for AWS S3
* `gs://<bucket>/<object>` for Google Cloud Storage

Additionally, the configuration format has been updated to include
sections on authentication tokens for the different services.

The authentication tokens are the query strings for presigned URLs.

This allows the engine configuration to store only a signature that
grants access to a resource without storing the storage account
keys.

As we aren't yet localizing inputs for local task executions, the
support for these URLs is currently limited to the `read_*`
functions from the WDL standard library; once input localization is
properly implemented, you'll be able to use these URIs in an inputs
file.