feat: incremental command line parsing
This commit started out as implementing incremental command line parsing.
Incremental command line parsing is as follows:

- Parse both files and key-value pairs on the command line and combine the
  results into a single input object.
- Overwrite earlier values provided on the command line arguments with command
  line arguments that come later.

This commit then ballooned a bit, as it was more fitting for some of those
pieces to be implemented over in the Sprocket code base
(https://github.com/stjude-rust-labs/sprocket). This then snowballed into a
larger discussion about moving functionality out of the `wdl` binary and into
`sprocket` proper, which I went ahead and knocked out in this commit.

Co-authored-by: Ari Frantz <ari.frantz@stjude.org>
Co-authored-by: Peter Huene <peter@huene.dev>