fix: path translation in expressions.
The fix in #417 was incomplete: there was more than just the `sep`
option that was failing to translate `File` and `Directory` paths.

This fixes the issue by moving the `write_primitive` function into
the `raw` display function for primitive values and allowing an
optional evaluation context to be passed that will be used to
translate paths.

The existing test has been updated to test more expressions.

Also corrected an unrelated test directory misspelling.