fix: the `sep` function is not translating paths.
The fix in #417 was incomplete: the `sep` function was also writing
the raw `File` and `Directory` values instead of doing path
translation.

This fixes the issue by extracting out the `write_primitive` helper
and calling it from the implementation of the `sep` function.

Also corrected an unrelated test directory misspelling.