feat: improve diagnostic for task execution failure.
The task execution failure message could be overly verbose and lacked links to
the source WDL.

This commit changes the error returned from evaluation to include a "callstack"
which is simply the locations of call statements; this can be used when
reporting errors to properly display a WDL evaluation callstack.

Additionally, the stderr output from the diagnostic was always printing _all_
of the stderr contents; this now effectively tails the last 10 lines of the
file in case the stderr stream was verbose.