fix: call locations spans should not span multiple lines.
This fixes call locations in a printed callstack displaying overly verbose
diagnostics due to using the entire span of the call statement; call statements
typically span many lines due to their inputs, leading to overly verbose
output.

The fix is to use the span of the call keyword in the call statement, which
guarantees only a single line of source output per call location.

This also includes a change to the evaluation tests so that invalid WDL gets a
failure message rather than writing to `errors.txt`. Additionally, the printed
output of the lint tests harness has been updated to match what we output for
the other test harnesses.