refactor: rename `crankshaft` backend to `docker`.
This commit renames the `crankshaft` backend to simply the `docker` backend and
removes the TODOs that were in place for the upcoming TES backend, which will
now be introduced as a separate backend given the number of differences between
the two backends (mostly due to how it handles inputs, outputs, and
localization).

Additionally, this updates `wdl-engine` to use `crankshaft` version `0.2.0`,
which required some changes in the (now) docker backend for handling stdout and
stderr.

Additionally, the backend configuration for the engine has subsequently changed
so that it no longer just wraps a Crankshaft configuration verbatim, allowing
for a stable configuration format driven by `wdl-engine`. The backend
configuration is now an enumeration rather than a set of structs, allowing for
a future where multiple backends might be used simultaneously or as fallbacks.

A number of the changes here are needed to support the upcoming TES backend as
well.