feat: add authentication to TES input and output URLs.
This commit adds authentication to TES input and output URLs.

Previously the backend worked with TES on Azure where the input and output URLs
did not need the authentication in the URLs because the TES server used its own
authentication.

To make the TES backend more generic, it should be sending the authentication
along with the inputs and outputs.