feat: allow multiple backends to be defined.
This commit changes the backend configuration in `wdl-engine` so
that multiple backends can be defined.

By default, it will look for a backend with the key `default` to
use as the default backend.

These changes will allow for better merging of configuration from
multiple sources and for users to switch between backends easily
by simply changing the `backend` configuration value to the name
of the backend to use.