refactor!: remove redundant inherent impls of `AstNode` methods
`Expr` and `PlaceholderOption` both had inherent impls of methods that have since been defined in
the `AstNode` trait. The implemenations are not completely identical, but the versions present in
the `AstNode` impls seem more robust, so this commit keeps those while removing the inherents. No
other changes were required tests continue to pass, so I expect this is a relative non-event despite
technically being a breaking change.