wip: recover previous check for unrecognized rules
The `Default` impl used to be a copy of `new()`, but with `unreachable!()` instead of `warn!()` in
the unrecognized rule case. This adds a panic when in the test configuration, so we still will have
an indication in test before shipping a rule that is not actually supported by `DiagnosticsConfig`.