fix: use the absolute path of an input file as its origin.
Parsing of an input file in `wdl-cli` was using the relative path to the file
as its origin.

As input files are commonly passed as relative paths on the CLI, this lead to
having relative paths for any file inputs it might contain.

However, for the initial inputs to a task or workflow, the paths should be
absolute.

The fix is to simply get the absolute path of the input file before getting its
parent directory, which is used as the origin path for any values the file
might contain.