refactor: refactors compound value type representation.
This commit refactors how compound values report their types so that they are
always unqualified.

Previously, compound values might store types which are qualified, meaning the
`?` or the `+` (in the case of an array type) are maintained.

However, this is inconsistent with how primitive values represent their type,
which is always unqualified.

This leads to issues where values might not coerce when they otherwise should.