fix: ensure `None` values have an associated type.
This commit fixes an issue where we were treating all `None` values as having
type `Type::None`. However, `Type::None` is the optional `Union` type which
allows type coercion to any other optional type.

For the purpose of common type calculations in array and map literals, this
coercion was too permissive. For example, an expression of `["foo", bar]` where
`bar` is of type `File?` should result in a calculated type of `Array[File?]`
and not `Array[String?]`. The former was the calculated type by `wdl-analysis`
and the latter the type calculated at evaluation time by `wdl-engine` because
the type of `bar` wasn't being associated with the resulting `None` value when
evaluated.

The fix keeps track of a type associated with `None` _values_. The type of
evaluating a literal `None` is always `Type::None`. However, that `None` value
can be coerced into another optional type and the target type then associated
with the resulting `None` value.

Fixes #548.