fix: fix deserialization of `Object` to not require WDL identifiers.
This commit loosens the restriction on `Object` deserialization such that the
keys are no longer required to be valid WDL identifiers.

This requirement was not in the WDL spec and prevented using the `Object` as an
intermediary that would be coerced to a different value type, such as `Map`,
for accessing the keys.

Additionally, this commit also implements support for type coercions between
`Map` and `Object`/`Struct` where the key type is coercible to and from
`String` instead of the previous requirement that the key type *be* `String`.

Fixes #549.