fix: correct array literal type calculation in `wdl-analysis`.
The type calculated for a non-empty array literal in `wdl-analysis` incorrectly
included a non-empty qualifier on the type.

This caused a panic in `wdl-engine` when an `if` expression was evaluated where one
side is a non-empty array and the other is an empty array (when the empty array
branch is taken). This is because the type of the non-empty branch included the
non-empty qualifier, but the branch taken was an empty array and the empty
array won't coerce to a non-empty qualified type.

As non-empty checks are a runtime constraint, `wdl-analysis` didn't catch that
the empty array would not coerce.

The fix is to simply drop the non-empty qualifier in the type calculation as it
should not have been included in the first place.

Fixes #578.