refactor: change how task inputs are evaluated.
Previously, `File` and `Directory` task inputs were evaluated and host paths
were translated to guest paths just prior to command evaluation.

However, private declarations of tasks may have already evaluated references to
the inputs which can cause host paths to end up in the evaluated command.

This commit refactors input evaluation so that the evaluation scope stores the
already mapped values of `File` and `Directory` inputs.

Additionally, input localization was also refactored out of the backends and
into task evaluation; backends no longer have to concern themselves with
mapping guest paths or downloading remote inputs.

The standard library functions dealing with reading and writing files had to be
updated to translate guest paths to host paths in the case of reading and host
paths to guest paths in case of writing.

The `InputTrie` was refactored to immediately calculate a guest path for a
newly inserted input (assuming the backend uses containers). Additionally, it
is not responsible for performing guest to host path translation.

Fixes #574