refactor: change how task inputs are evaluated.
Previously, `File` and `Directory` task inputs were evaluated and host paths
were translated to guest paths just prior to command evaluation.

However, private declarations of tasks may have already evaluated references to
the inputs which can cause host paths to end up in the evaluated command.

To address this, task evaluation has been refactored to calculate guest paths
upon evaluation of an input, evaluation of a private declaration, or the
creation of a temporary file from a stdlib function.

When a `File` or `Directory` is coerced to a `String` or when it is
interpolated, guest paths are substituted for host paths. When a `String` is
coerced to `File` or `Directory` and it is exactly a known guest path, a host
path is substituted.

Additionally, backends are no longer concerned with mapping paths or with
localizing inputs; that now occurs during task evaluation and the backends
receive inputs with host, guest, and localized paths.

The following was also fixed as a part of these changes:

* The `InputTrie` data structure was moved into its own module.
* Non-optional `File` and `Directory` inputs are validated for existence for
  WDL 1.2 documents.
* Relative paths in WDL documents are now treated as relative to the containing
  document; a "base directory" is now more consistently used through the
  source.
* The `join_paths` function now properly handles relative paths (relative to
  the evaluation's base directory).
* Temporary file inputs are now sent to the backend individually instead of as
  a `Directory` input of the evaluation's `tmp` directory itself.

Fixes #574