feat: use `cloud-copy` for file transfers.
This commit replaces the logic in the `http` module with an implementation
based on `cloud-copy` to support file uploads to Azure Blob Storage, AWS S3,
and Google Cloud Storage (GCS).

The configuration for S3 and GCS are now based on account keys and secrets as
presigning URLs requires the URL be known ahead of time (as opposed to Azure's
SAS tokens which can operate at the container level; Azure's authentication
configuration remains unchanged).

Additionally, file uploading is now implemented for the TES backend.

The `Downloader` trait has been renamed to `Transferer` as it now handles
uploads in addition to downloads.

Transfer events can now be emitted from evaluation, so an `Events` type that
handles the creation of the broadcast channels was added.

The cloud service URL rewriting implementation was removed in favor of using
the one from `cloud-copy`.

A cryptographic hashing implementation using `blake3` was added so that the TES
backend will only upload content it hasn't uploaded before (note: integrity of
the remote data is not currently validated).