fix: only add a backend input for specified inputs.
Fix check in the Docker backend that requires an input to exist for it to be
included in the bind mounts.

The underlying problem is that a backend input should only be added for a
specified input as it may contain host or remote paths; for inputs that are
defaulted, the evaluated expression will only contain guest paths and therefore
does not need an input.