fix: ensure correct directory permissions for work directory.
This commit fixes the Docker backend such that it creates the work directory
with `770` permissions so that the working directory is writable to the user
used by the container.

Also, this fixes backend cleanup to occur just before task output evaluation so
that the current user can read from the work directory regardless of what user
the container used.