fix: correct assertion for TES backend.
The changes in #605 made the assertion of the work directory being local in
the wrong place, which impacted the TES backend and caused a panic.

The fix is to take cleanup by an `EvaluationPath` which may be local or remote;
the assertion was moved to the Docker backend where the path should always be
local.